# contracts/contract3-api.openapi.yaml
openapi: 3.1.0
info:
  title: DarkPulse Investigator API
  version: 1.0.0
paths:
...
  /slang:
    get:
      tags:
      - Slang
      summary: List Slang
      operationId: list_slang_api_v1_slang_get
      security:
      - HTTPBearer: []
      parameters:
      - name: lang
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Lang
      - name: newly_discovered
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Newly Discovered
      - name: review_status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Review Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlangListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Slang
      summary: Create Slang Entry
      operationId: create_slang_entry_api_v1_slang_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SlangEntry'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /slang/candidates:
    get:
      tags:
      - Slang
      summary: List Candidates
      operationId: list_candidates_api_v1_slang_candidates_get
      security:
      - HTTPBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlangListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /slang/{slang_id}:
    put:
      tags:
      - Slang
      summary: Update Slang Entry
      operationId: update_slang_entry_api_v1_slang__slang_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: slang_id
        in: path
        required: true
        schema:
          type: string
          title: Slang Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SlangUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Slang
      summary: Delete Slang Entry
      operationId: delete_slang_entry_api_v1_slang__slang_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: slang_id
        in: path
        required: true
        schema:
          type: string
          title: Slang Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /slang/{slang_id}/approve:
    post:
      tags:
      - Slang
      summary: Approve Candidate
      operationId: approve_candidate_api_v1_slang__slang_id__approve_post
      security:
      - HTTPBearer: []
      parameters:
      - name: slang_id
        in: path
        required: true
        schema:
          type: string
          title: Slang Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /slang/{slang_id}/reject:
    post:
      tags:
      - Slang
      summary: Reject Candidate
      operationId: reject_candidate_api_v1_slang__slang_id__reject_post
      security:
      - HTTPBearer: []
      parameters:
      - name: slang_id
        in: path
        required: true
        schema:
          type: string
          title: Slang Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /operations/sources:
    get:
      tags:
      - Operations
      summary: List Sources
      operationId: list_sources_api_v1_operations_sources_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      security:
      - HTTPBearer: []
  /operations/processing:
    get:
      tags:
      - Operations
      summary: Processing Status
      operationId: processing_status_api_v1_operations_processing_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      security:
      - HTTPBearer: []
  /operations/onion-review:
    get:
      tags:
      - Operations
      summary: Onion Review Status
      operationId: onion_review_status_api_v1_operations_onion_review_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      security:
      - HTTPBearer: []
  /operations/audit:
    get:
      tags:
      - Operations
      summary: List Operations Audit
      operationId: list_operations_audit_api_v1_operations_audit_get
      security:
      - HTTPBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /operations/collection-runs:
    get:
      tags:
      - Operations
      summary: List Collection Runs
      operationId: list_collection_runs_api_v1_operations_collection_runs_get
      security:
      - HTTPBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /operations/dead-letter/{ingest_id}/retry:
    post:
      tags:
      - Operations
      summary: Retry Dead Letter
      operationId: retry_dead_letter_api_v1_operations_dead_letter__ingest_id__retry_post
      security:
      - HTTPBearer: []
      parameters:
      - name: ingest_id
        in: path
        required: true
        schema:
          type: string
          title: Ingest Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApiEnvelope:
      properties:
        data:
          title: Data
        pagination:
          anyOf:
          - $ref: '#/components/schemas/Pagination'
          - type: 'null'
        meta:
          additionalProperties: true
          type: object
          title: Meta
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          type: array
          title: Errors
      type: object
      title: ApiEnvelope
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Trace Id
      type: object
      required:
      - code
      - message
      title: ErrorDetail
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Pagination:
      properties:
        cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        limit:
          type: integer
          title: Limit
          default: 50
        total:
          type: integer
          title: Total
          default: 0
      type: object
      title: Pagination
    SlangEntry:
      properties:
        term:
          type: string
          maxLength: 200
          minLength: 1
          title: Term
        meaning:
          type: string
          maxLength: 200
          minLength: 1
          title: Meaning
        lang:
          type: string
          maxLength: 32
          minLength: 1
          title: Lang
          default: en
        confidence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Confidence
          default: 1.0
        newly_discovered:
          type: boolean
          title: Newly Discovered
          default: false
      type: object
      required:
      - term
      - meaning
      title: SlangEntry
    SlangListResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/SlangResponse'
          type: array
          title: Data
        pagination:
          anyOf:
          - $ref: '#/components/schemas/Pagination'
          - type: 'null'
        meta:
          additionalProperties: true
          type: object
          title: Meta
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          type: array
          title: Errors
      type: object
      title: SlangListResponse
    SlangResponse:
      properties:
        term:
          type: string
          maxLength: 200
          minLength: 1
          title: Term
        meaning:
          type: string
          maxLength: 200
          minLength: 1
          title: Meaning
        lang:
          type: string
          maxLength: 32
          minLength: 1
          title: Lang
          default: en
        confidence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Confidence
          default: 1.0
        newly_discovered:
          type: boolean
          title: Newly Discovered
          default: false
        review_status:
          type: string
          enum:
          - pending
          - approved
          - rejected
          title: Review Status
          default: approved
        id:
          type: string
          title: Id
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
        usage_count:
          type: integer
          title: Usage Count
          default: 0
      type: object
      required:
      - term
      - meaning
      - id
      title: SlangResponse
    SlangUpdate:
      properties:
        term:
          type: string
          maxLength: 200
          minLength: 1
          title: Term
        meaning:
          type: string
          maxLength: 200
          minLength: 1
          title: Meaning
        lang:
          type: string
          maxLength: 32
          minLength: 1
          title: Lang
          default: en
        confidence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Confidence
          default: 1.0
        newly_discovered:
          type: boolean
          title: Newly Discovered
          default: false
        review_status:
          type: string
          enum:
          - pending
          - approved
          - rejected
          title: Review Status
          default: approved
      type: object
      required:
      - term
      - meaning
      title: SlangUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: http://localhost:8003